Move `RegisterView` to accounts.views
